base_datos_gru_semana/requirements.txt
pandas>=2.1
numpy>=1.26
sqlalchemy>=2.0
pyodbc>=5.1
python-dotenv>=1.0
tqdm>=4.66
